refactor(rss): move RSS helpers to shared utility module
